# wikiwatch-backend/WikiWatch-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WikiWatch API",
    "description": "Instructor-provided backend for frontend training. One shared team. Real database, role permissions, review transitions and plain-text diff comments. Public Wikimedia reads stay in the browser. Login, copy access_token, then use Authorize. All JSON success responses use {success, data}.",
    "version": "1.0.0"
  },
  "paths": {
    "/wikiwatch-service/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in with email and password",
        "operationId": "login_wikiwatch_service_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_Tokens_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      }
    },
    "/wikiwatch-service/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate refresh and access tokens",
        "description": "Each refresh token can be used once. Old access token is revoked too. Serialize refresh requests in the frontend.",
        "operationId": "refresh_wikiwatch_service_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_Tokens_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      }
    },
    "/wikiwatch-service/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke the current session",
        "operationId": "logout_wikiwatch_service_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_LogoutResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/wikiwatch-service/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Read the current account and role",
        "operationId": "me_wikiwatch_service_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MemberResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/wikiwatch-service/v1/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "List team members (lead or admin)",
        "operationId": "members_wikiwatch_service_v1_members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_Page_MemberResponse__"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Create a member account (admin)",
        "description": "Create an account with a password shared separately by the trainer. This does not send an invitation email. No public signup.",
        "operationId": "create_wikiwatch_service_v1_members_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MemberResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      }
    },
    "/wikiwatch-service/v1/members/{member_id}": {
      "patch": {
        "tags": [
          "Members"
        ],
        "summary": "Change role or active status (admin)",
        "description": "Cannot remove your own admin access. Unfinished claims must be reassigned or completed before reviewer access is removed.",
        "operationId": "change_wikiwatch_service_v1_members__member_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MemberResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      }
    },
    "/wikiwatch-service/v1/members/directory": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Read names for comment authors and claim owners",
        "description": "All roles. This directory excludes email addresses and password hashes.",
        "operationId": "directory_wikiwatch_service_v1_members_directory_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_Page_DirectoryMember__"
                }
              }
            }
          },
          "401": {
            "description": "Session expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role or ownership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict: reload before retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or transition reason"
          },
          "429": {
            "description": "Login rate limit reached"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DirectoryMember": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "active"
        ],
        "title": "DirectoryMember"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LogoutResponse": {
        "properties": {
          "logged_out": {
            "type": "boolean",
            "const": true,
            "title": "Logged Out",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LogoutResponse"
      },
      "MemberCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email",
          "role",
          "password"
        ],
        "title": "MemberCreate"
      },
      "MemberResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "active"
        ],
        "title": "MemberResponse"
      },
      "MemberUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MemberUpdate"
      },
...
      "Page_DirectoryMember_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DirectoryMember"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[DirectoryMember]"
      },
...
      "Page_MemberResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "Page[MemberResponse]"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "format": "password",
            "title": "Refresh Token",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "reviewer",
          "lead",
          "admin"
        ],
        "title": "Role"
      },
...
      "SuccessResponse_LogoutResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/LogoutResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[LogoutResponse]"
      },
      "SuccessResponse_MemberResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[MemberResponse]"
      },
...
      "SuccessResponse_Page_DirectoryMember__": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/Page_DirectoryMember_"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[Page[DirectoryMember]]"
      },
...
      "SuccessResponse_Page_MemberResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/Page_MemberResponse_"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[Page[MemberResponse]]"
      },
...
      "SuccessResponse_Tokens_": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/Tokens"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[Tokens]"
      },
...
      "Tokens": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "member": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "member"
        ],
        "title": "Tokens"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Paste the access_token returned by login. Do not use the refresh token.",
        "scheme": "bearer"
      }
    }
  }
}